social_django/requirements.txt
anyio==3.6.2
asgiref==3.6.0
asttokens==2.2.1
backcall==0.2.0
certifi==2023.7.22
charset-normalizer==3.3.1
click==8.1.3
colorama==0.4.6
decorator==5.1.1
distlib==0.3.6
Django==4.2.4
django-enumchoicefield==3.0.1
django-simple-history==3.4.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.0
executing==1.2.0
factory-boy==3.2.1
Faker==18.7.0
fastapi==0.95.0
filelock==3.12.2
h11==0.14.0
httptools==0.5.0
idna==3.4
iniconfig==2.0.0
ipdb==0.13.13
ipython==8.12.0
jedi==0.18.2
matplotlib-inline==0.1.6
numpy==1.24.2
packaging==23.1
pandas==2.0.0
parso==0.8.3
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.8.0
pluggy==1.0.0
prompt-toolkit==3.0.38
pure-eval==0.2.2
pydantic==1.10.7
Pygments==2.15.1
PySimpleGUI==4.60.4
pytest==7.3.1
pytest-django==4.5.2
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0
requests==2.31.0
selenium==4.24.0
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.4
stack-data==0.6.2
starlette==0.26.1
traitlets==5.9.0
typing_extensions==4.5.0
tzdata==2023.3
urllib3==2.0.7
uvicorn==0.21.1
virtualenv==20.23.1
watchfiles==0.19.0
wcwidth==0.2.6
websockets==10.4
